# backend/api/spec/openapi-v1.yaml
openapi: 3.1.0
info:
  title: AetherFlow API
  version: 3.1.0
  description: Canonical schema for `/api/v1`; legacy `/api` remains available during migration.
servers:
  - { url: /api/v1, description: Canonical v1 }
  - { url: /api, description: Legacy compatibility }
...
paths:
...
  /ws: { get: { tags: [System], summary: Metrics websocket, responses: { '101': { description: Switching Protocols }, '401': { $ref: '#/components/responses/Error' } } } }
  /ws/logs: { get: { tags: [System], summary: Logs websocket, responses: { '101': { description: Switching Protocols }, '401': { $ref: '#/components/responses/Error' } } } }
...
  /auth/login: { post: { tags: [Auth], summary: Login with username and password, requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { type: object, properties: { message: { type: string }, user: { $ref: '#/components/schemas/User' } } } } } }, '400': { $ref: '#/components/responses/Error' }, '401': { $ref: '#/components/responses/Error' } } } }
  /auth/setup: { post: { tags: [Auth], summary: Create initial admin, requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } }, '409': { $ref: '#/components/responses/Error' } } } }
...
  /auth/session: { get: { tags: [Auth], summary: Get current session, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }, '401': { $ref: '#/components/responses/Error' } } } }
...
  /auth/profile: { put: { tags: [Auth], summary: Update current profile, security: [ { cookieAuth: [] } ], requestBody: { required: true, content: { application/json: { schema: { type: object, required: [email], properties: { email: { type: string } } } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } } } } }
  /settings:
    get: { tags: [System], summary: Get settings, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Settings' } } } } } }
    put: { tags: [System], summary: Update settings, security: [ { cookieAuth: [] } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Settings' } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { type: object, properties: { message: { type: string }, data: { $ref: '#/components/schemas/Settings' } } } } } } } }
  /settings/test-ai: { post: { tags: [AI], summary: Test AI provider, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
...
  /system/update/run: { post: { tags: [System], summary: Run release update, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } } } } }
...
  /services/{name}/control: { post: { tags: [Services], summary: Control a service, security: [ { cookieAuth: [] } ], parameters: [ { $ref: '#/components/parameters/ServiceName' } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ServiceControlRequest' } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
...
  /packages/{id}/install: { post: { tags: [Marketplace], summary: Install a package, security: [ { cookieAuth: [] } ], parameters: [ { $ref: '#/components/parameters/PackageID' } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /packages/{id}/uninstall: { post: { tags: [Marketplace], summary: Uninstall a package, security: [ { cookieAuth: [] } ], parameters: [ { $ref: '#/components/parameters/PackageID' } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /packages/{id}/progress: { get: { tags: [Marketplace], summary: Package progress, parameters: [ { $ref: '#/components/parameters/PackageID' } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PackageProgress' } } } } } } }
...
  /fileshare/upload: { post: { tags: [Files], summary: Upload a file, security: [ { cookieAuth: [] } ], requestBody: { required: true, content: { multipart/form-data: { schema: { type: object, required: [file], properties: { file: { type: string, format: binary } } } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } }, '507': { $ref: '#/components/responses/Error' } } } }
  /fileshare/download/{filename}: { get: { tags: [Files], summary: Download a fileshare item, security: [ { cookieAuth: [] } ], parameters: [ { $ref: '#/components/parameters/Filename' } ], responses: { '200': { description: File stream, content: { application/octet-stream: { schema: { type: string, format: binary } } } } } } }
  /backup/run: { post: { tags: [Backup], summary: Run backup, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /backup/list: { get: { tags: [Backup], summary: List backups, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericArray' } } } } } } }
  /backup/download/{filename}: { get: { tags: [Backup], summary: Download backup, security: [ { cookieAuth: [] } ], parameters: [ { $ref: '#/components/parameters/Filename' } ], responses: { '200': { description: Backup stream, content: { application/octet-stream: { schema: { type: string, format: binary } } } } } } }
  /backup/upload/{filename}: { post: { tags: [Backup], summary: Upload backup chunk, security: [ { cookieAuth: [] } ], parameters: [ { $ref: '#/components/parameters/Filename' } ], requestBody: { required: true, content: { application/octet-stream: { schema: { type: string, format: binary } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /users: { get: { tags: [Users], summary: List users, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/User' } } } } } } } }
  /users/{id}/role: { put: { tags: [Users], summary: Update user role, security: [ { cookieAuth: [] } ], parameters: [ { $ref: '#/components/parameters/UserID' } ], requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { role: { type: string } }, required: [role] } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } } } } }
  /users/{id}: { delete: { tags: [Users], summary: Delete user, security: [ { cookieAuth: [] } ], parameters: [ { $ref: '#/components/parameters/UserID' } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } } } } }
  /user/quota: { get: { tags: [Quotas], summary: Get current user quota, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/UserQuota' } } } } } } }
  /quotas: { get: { tags: [Quotas], summary: List quotas, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { type: object, properties: { quotas: { type: array, items: { $ref: '#/components/schemas/UserQuota' } } } } } } } } } }
  /quotas/{id}: { put: { tags: [Quotas], summary: Set quota, security: [ { cookieAuth: [] } ], parameters: [ { $ref: '#/components/parameters/UserID' } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { type: object, properties: { message: { type: string }, quota: { $ref: '#/components/schemas/UserQuota' } } } } } } } } }
  /quotas/{id}/refresh: { post: { tags: [Quotas], summary: Refresh quota usage, security: [ { cookieAuth: [] } ], parameters: [ { $ref: '#/components/parameters/UserID' } ], responses: { '200': { description: OK, content: { application/json: { schema: { type: object, properties: { quota: { $ref: '#/components/schemas/UserQuota' } } } } } } } } }
  /billing/webhooks: { get: { tags: [Billing], summary: List billing webhook audits, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { type: object, properties: { events: { type: array, items: { $ref: '#/components/schemas/BillingWebhookAudit' } } } } } } } } } }
  /billing/webhooks/{provider}: { post: { tags: [Billing], summary: Receive billing webhook, parameters: [ { in: path, name: provider, required: true, schema: { type: string, enum: [whmcs, blesta] } } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/BillingWebhookResult' } } } }, '400': { $ref: '#/components/responses/Error' }, '503': { $ref: '#/components/responses/Error' } } } }
...
  /notifications/{id}/read: { put: { tags: [Notifications], summary: Mark notification as read, parameters: [ { in: path, name: id, required: true, schema: { type: string } } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } } } } }
...
  /notifications/rules: { get: { tags: [Notifications, Operator], summary: List notification rules, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } }, post: { tags: [Notifications, Operator], summary: Create notification rule, security: [ { cookieAuth: [] } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } }, responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /notifications/rules/{id}: { put: { tags: [Notifications, Operator], summary: Update notification rule, security: [ { cookieAuth: [] } ], parameters: [ { in: path, name: id, required: true, schema: { type: string } } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } } } }, delete: { tags: [Notifications, Operator], summary: Delete notification rule, security: [ { cookieAuth: [] } ], parameters: [ { in: path, name: id, required: true, schema: { type: string } } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } } } } }
  /notifications/channels: { get: { tags: [Notifications, Operator], summary: List notification channels, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } }, post: { tags: [Notifications, Operator], summary: Create notification channel, security: [ { cookieAuth: [] } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } }, responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /notifications/channels/{id}/test: { post: { tags: [Notifications, Operator], summary: Test notification channel, security: [ { cookieAuth: [] } ], parameters: [ { in: path, name: id, required: true, schema: { type: string } } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } } } } }
  /notifications/channels/{id}: { delete: { tags: [Notifications, Operator], summary: Delete notification channel, security: [ { cookieAuth: [] } ], parameters: [ { in: path, name: id, required: true, schema: { type: string } } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } } } } }
  /cluster/nodes: { get: { tags: [Cluster, Operator], summary: List cluster nodes, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericArray' } } } } } }, post: { tags: [Cluster, Operator], summary: Enroll worker, security: [ { cookieAuth: [] } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /cluster/nodes/{id}: { delete: { tags: [Cluster, Operator], summary: Remove worker, security: [ { cookieAuth: [] } ], parameters: [ { $ref: '#/components/parameters/UserID' } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } } } } }
...
  /oidc/clients: { get: { tags: [OIDC, Operator], summary: List OIDC clients, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericArray' } } } } } }, post: { tags: [OIDC, Operator], summary: Create OIDC client, security: [ { cookieAuth: [] } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } }, responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /oidc/clients/{id}: { delete: { tags: [OIDC, Operator], summary: Delete OIDC client, security: [ { cookieAuth: [] } ], parameters: [ { in: path, name: id, required: true, schema: { type: string } } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } } } } }
...
  /oidc/token: { post: { tags: [OIDC], summary: OIDC token endpoint, requestBody: { required: true, content: { application/x-www-form-urlencoded: { schema: { $ref: '#/components/schemas/GenericObject' } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
...
  /oidc/revoke: { post: { tags: [OIDC], summary: OIDC revoke endpoint, requestBody: { required: true, content: { application/x-www-form-urlencoded: { schema: { $ref: '#/components/schemas/GenericObject' } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { type: object } } } } } } }
  /ai/chat: { post: { tags: [AI], summary: Chat with FlowAI, requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /ai/support: { post: { tags: [AI], summary: AI support helper, requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /ai/metadata/scan: { post: { tags: [AI], summary: Start metadata scan, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /ai/metadata/status: { get: { tags: [AI], summary: Metadata scan status, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /ai/metadata/results: { get: { tags: [AI], summary: Metadata scan results, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericArray' } } } } } } }
  /ai/bandwidth/analyze: { post: { tags: [AI], summary: Analyze bandwidth, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /ai/bandwidth/apply: { post: { tags: [AI], summary: Apply bandwidth plan, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /ai/predictions: { get: { tags: [AI, Operator], summary: Predictive scaling data, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /ai/predictions/analyze: { post: { tags: [AI, Operator], summary: Run predictive analysis, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /ai/predictions/history: { get: { tags: [AI, Operator], summary: Predictive history, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericArray' } } } } } } }
  /ai/backup/optimal-window: { get: { tags: [AI], summary: Recommended backup window, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /ai/backup/schedule: { post: { tags: [AI], summary: Set backup schedule, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /logs: { get: { tags: [System], summary: Query logs, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericArray' } } } } } } }
  /logs/sources: { get: { tags: [System], summary: List log sources, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericArray' } } } } } } }
  /logs/bookmarks:
    get: { tags: [System], summary: List saved log bookmarks, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } }
    post: { tags: [System], summary: Bookmark log entry, security: [ { cookieAuth: [] } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } } } }
  /network/status: { get: { tags: [Network, Operator], summary: Network status, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /network/wireguard/peers: { get: { tags: [Network, Operator], summary: List WireGuard peers, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericArray' } } } } } }, post: { tags: [Network, Operator], summary: Add WireGuard peer, security: [ { cookieAuth: [] } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /network/wireguard/peers/{key}: { delete: { tags: [Network, Operator], summary: Remove WireGuard peer, security: [ { cookieAuth: [] } ], parameters: [ { in: path, name: key, required: true, schema: { type: string } } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MessageResponse' } } } } } } }
  /network/wireguard/keygen: { post: { tags: [Network, Operator], summary: Generate WireGuard keys, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /network/tailscale/status: { get: { tags: [Network, Operator], summary: Tailscale status, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
  /network/tailscale/peers: { get: { tags: [Network, Operator], summary: List Tailscale peers, security: [ { cookieAuth: [] } ], responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericArray' } } } } } } }
  /network/tailscale/routes: { post: { tags: [Network, Operator], summary: Advertise Tailscale routes, security: [ { cookieAuth: [] } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } }, responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/GenericObject' } } } } } } }
components:
  securitySchemes: { cookieAuth: { type: apiKey, in: cookie, name: aetherflow_session } }
  parameters:
    UserID: { in: path, name: id, required: true, schema: { type: integer } }
    PackageID: { in: path, name: id, required: true, schema: { type: string } }
    ServiceName: { in: path, name: name, required: true, schema: { type: string } }
    Filename: { in: path, name: filename, required: true, schema: { type: string } }
  responses:
    Error: { description: Error response, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  schemas:
    ErrorResponse: { type: object, required: [error], properties: { error: { type: string } } }
    MessageResponse: { type: object, properties: { message: { type: string } } }
    LoginRequest: { type: object, required: [username, password], properties: { username: { type: string }, password: { type: string } } }
    User: { type: object, properties: { id: { type: integer }, username: { type: string }, email: { type: string }, avatar_url: { type: string }, role: { type: string }, is_oauth: { type: boolean } } }
    Settings: { type: object, properties: { aiModel: { type: string }, systemPrompt: { type: string }, language: { type: string }, timezone: { type: string }, updateChannel: { type: string }, defaultDashboard: { type: string }, setupCompleted: { type: boolean }, geminiApiKey: { type: string } } }
    ServiceControlRequest: { type: object, required: [action], properties: { action: { type: string, enum: [start, stop, restart] }, managed_by: { type: string }, process: { type: string } } }
...
    PackageProgress: { type: object, properties: { status: { type: string }, progress: { type: integer }, started_at: { type: string }, log_line: { type: string }, log_lines: { type: integer } } }
...
    UserQuota: { type: object, properties: { user_id: { type: integer }, username: { type: string }, quota_bytes: { type: integer }, used_bytes: { type: integer }, available_bytes: { type: integer }, usage_pct: { type: number }, configured: { type: boolean }, status: { type: string }, source: { type: string }, billing_provider: { type: string }, billing_external_id: { type: string }, updated_at: { type: string } } }
    BillingWebhookAudit: { type: object, properties: { id: { type: integer }, provider: { type: string }, event_id: { type: string }, event_type: { type: string }, username: { type: string }, quota_bytes: { type: integer }, status: { type: string }, error: { type: string }, payload: { type: string }, processed_at: { type: string } } }
    BillingWebhookResult: { type: object, properties: { provider: { type: string }, event_id: { type: string }, event_type: { type: string }, username: { type: string }, quota_bytes: { type: integer }, applied: { type: boolean }, status: { type: string }, message: { type: string }, quota: { $ref: '#/components/schemas/UserQuota' } } }
    GenericObject: { type: object, additionalProperties: true }
    GenericArray: { type: array, items: { type: object, additionalProperties: true } }
...